refactor(web): renames StackSizes component into Indicators
