+++ apps/web/src/components/Indicators.svelte
@@ -13,8 +13,8 @@
   }
 </style>
 
-{#each items as { screenPosition: { x, y }, size, id } (id)}
+{#each items as { screenPosition: { x, y }, size, name, id } (id)}
   <div class="indicator" style="top: {y}px; left: {x}px">
-    {size}
+    {name ?? size}
   </div>
 {/each}
