+++ apps/web/src/components/index.js
@@ -9,6 +9,7 @@ export { default as Discussion } from './Discussion.svelte'
 export { default as Dropdown } from './Dropdown.svelte'
 export { default as GameAside } from './GameAside.svelte'
 export { default as GameLink } from './GameLink.svelte'
+export { default as Indicators } from './Indicators.svelte'
 export { default as Input } from './Input.svelte'
 export { default as MeshDetails } from './MeshDetails.svelte'
 export { default as Menu } from './Menu.svelte'
@@ -19,5 +20,4 @@ export { default as Progress } from './Progress.svelte'
 export { default as QuantityButton } from './QuantityButton.svelte'
 export { default as RadialMenu } from './RadialMenu.svelte'
 export { default as RuleViewer } from './RuleViewer.svelte'
-export { default as StackSizes } from './StackSizes.svelte'
 export { default as Typeahead } from './Typeahead.svelte'
